chore(refactor): separate `openvm-continuations` from SDK
