+++ benchmarks/src/bin/fib_e2e.rs
@@ -10,7 +10,8 @@ use openvm_rv32im_transpiler::{
     Rv32ITranspilerExtension, Rv32IoTranspilerExtension, Rv32MTranspilerExtension,
 };
 use openvm_sdk::{
-    commit::commit_app_exe, keygen::RootVerifierProvingKey, prover::ContinuationProver, Sdk, StdIn,
+    commit::commit_app_exe, keygen::RootVerifierProvingKey, prover::ContinuationProver,
+    DefaultStaticVerifierPvHandler, Sdk, StdIn,
 };
 use openvm_stark_sdk::bench::run_with_metric_collection;
 use openvm_transpiler::{transpiler::Transpiler, FromElf};
@@ -40,7 +41,7 @@ async fn main() -> Result<()> {
     let full_agg_pk = sdk.agg_keygen(
         agg_config,
         &halo2_params_reader,
-        None::<&RootVerifierProvingKey>,
+        &DefaultStaticVerifierPvHandler,
     )?;
     let elf = args.build_bench_program("fibonacci")?;
     let exe = VmExe::from_elf(
