+++ crates/continuations/src/verifier/common/types.rs
