+++ crates/continuations/src/verifier/root/vars.rs
