+++ crates/sdk/src/config/mod.rs
@@ -1,5 +1,9 @@
 use openvm_circuit::arch::instructions::program::DEFAULT_MAX_NUM_PUBLIC_VALUES;
-use openvm_native_compiler::conversion::CompilerOptions;
+use openvm_continuations::verifier::{
+    common::types::VmVerifierPvs, internal::types::InternalVmVerifierPvs,
+};
+use openvm_native_circuit::NativeConfig;
+use openvm_native_compiler::{conversion::CompilerOptions, ir::DIGEST_SIZE};
 use openvm_stark_sdk::config::FriParameters;
 use serde::{Deserialize, Serialize};
 
@@ -151,5 +155,35 @@ impl Default for LeafFriParams {
 impl From<FriParameters> for LeafFriParams {
     fn from(fri_params: FriParameters) -> Self {
         Self { fri_params }
+    }
+}
+
+const SBOX_SIZE: usize = 7;
+
+impl AggStarkConfig {
+    pub fn leaf_vm_config(&self) -> NativeConfig {
+        let mut config = NativeConfig::aggregation(
+            VmVerifierPvs::<u8>::width(),
+            SBOX_SIZE.min(self.leaf_fri_params.max_constraint_degree()),
+        );
+        config.system.profiling = self.profiling;
+        config
+    }
+    pub fn internal_vm_config(&self) -> NativeConfig {
+        let mut config = NativeConfig::aggregation(
+            InternalVmVerifierPvs::<u8>::width(),
+            SBOX_SIZE.min(self.internal_fri_params.max_constraint_degree()),
+        );
+        config.system.profiling = self.profiling;
+        config
+    }
+    pub fn root_verifier_vm_config(&self) -> NativeConfig {
+        let mut config = NativeConfig::aggregation(
+            // app_commit + leaf_verifier_commit + public_values
+            DIGEST_SIZE * 2 + self.max_num_user_public_values,
+            SBOX_SIZE.min(self.root_fri_params.max_constraint_degree()),
+        );
+        config.system.profiling = self.profiling;
+        config
     }
 }
