+++ crates/sdk/src/keygen/perm.rs
@@ -1,8 +1,7 @@
 use std::cmp::Reverse;
 
 use openvm_circuit::arch::{CONNECTOR_AIR_ID, PROGRAM_AIR_ID, PUBLIC_VALUES_AIR_ID};
-
-use crate::verifier::common::types::SpecialAirIds;
+use openvm_continuations::verifier::common::types::SpecialAirIds;
 
 pub struct AirIdPermutation {
     pub perm: Vec<usize>,
