+++ crates/sdk/src/prover/app.rs
@@ -1,16 +1,14 @@
 use std::sync::Arc;
 
 use getset::Getters;
-use openvm_circuit::arch::VmConfig;
+use openvm_circuit::arch::{ContinuationVmProof, VmConfig};
 use openvm_stark_backend::{proof::Proof, Chip};
 use openvm_stark_sdk::config::baby_bear_poseidon2::BabyBearPoseidon2Engine;
 use tracing::info_span;
 
 use super::vm::SingleSegmentVmProver;
 use crate::{
-    prover::vm::{
-        local::VmLocalProver, types::VmProvingKey, ContinuationVmProof, ContinuationVmProver,
-    },
+    prover::vm::{local::VmLocalProver, types::VmProvingKey, ContinuationVmProver},
     NonRootCommittedExe, StdIn, F, SC,
 };
