+++ crates/sdk/src/prover/root.rs
@@ -1,5 +1,6 @@
 use async_trait::async_trait;
 use openvm_circuit::arch::{SingleSegmentVmExecutor, Streams};
+use openvm_continuations::verifier::root::types::RootVmVerifierInput;
 use openvm_native_circuit::NativeConfig;
 use openvm_native_recursion::hints::Hintable;
 use openvm_stark_sdk::{
@@ -11,7 +12,6 @@ use openvm_stark_sdk::{
 use crate::{
     keygen::RootVerifierProvingKey,
     prover::vm::{AsyncSingleSegmentVmProver, SingleSegmentVmProver},
-    verifier::root::types::RootVmVerifierInput,
     RootSC, F, SC,
 };
