+++ crates/sdk/src/prover/vm/mod.rs
@@ -1,28 +1,12 @@
 use async_trait::async_trait;
-use derivative::Derivative;
-use openvm_circuit::{
-    arch::Streams,
-    system::memory::{tree::public_values::UserPublicValuesProof, CHUNK},
-};
+use openvm_circuit::arch::{ContinuationVmProof, Streams};
 use openvm_stark_backend::{
-    config::{Com, StarkGenericConfig, Val},
+    config::{StarkGenericConfig, Val},
     proof::Proof,
 };
-use serde::{Deserialize, Serialize};
 
 pub mod local;
 pub mod types;
-
-#[derive(Serialize, Deserialize, Derivative)]
-#[derivative(Clone(bound = "Com<SC>: Clone"))]
-#[serde(bound(
-    serialize = "Com<SC>: Serialize",
-    deserialize = "Com<SC>: Deserialize<'de>"
-))]
-pub struct ContinuationVmProof<SC: StarkGenericConfig> {
-    pub per_segment: Vec<Proof<SC>>,
-    pub user_public_values: UserPublicValuesProof<{ CHUNK }, Val<SC>>,
-}
 
 /// Prover for a specific exe in a specific continuation VM using a specific Stark config.
 pub trait ContinuationVmProver<SC: StarkGenericConfig> {
